(Header): Add SidebarButton
Refactor of Header values to choose appropriate according to scroll.
Add  a getAccordingScrollValue to increase readable and reduce lines.
Add SidebarButton on Header with props like onlySmall and onlyBig to show in some devices, also allow you to use a function to render your own sidebar button and use the internal header state to know if it's open and a function to change his state.